Converted the apio upload command to the new proto based scons params. (ice40 only).
